Helper Functions/requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
anyio==4.7.0
astroid==3.3.7
async-timeout==5.0.1
attrs==24.3.0
backtrader==1.9.78.123
base58==2.1.1
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
cachetools==4.2.4
ccxt==4.4.42
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cryptography==44.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.2.0
dill==0.3.9
exceptiongroup==1.2.2
filelock==3.17.0
flake8==7.1.1
Flask==2.2.5
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonalias==0.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
narwhals==1.18.4
networkx==3.4.2
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==3.20.3
pyaml==24.12.1
pyarrow==18.1.0
pycares==4.5.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.3.2
PyMySQL==1.1.1
pyparsing==3.2.1
pyserum==0.5.0a0
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ta==2.9.1
python-telegram-bot==21.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
schedule==1.2.2
scikit-learn==1.6.0
scikit-optimize==0.10.2
scipy==1.14.1
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
solana==0.36.2
solders==0.23.0
soupsieve==2.6
streamlit==1.41.1
sympy==1.13.1
ta==0.11.0
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
typeguard==4.4.1
types-cachetools==4.2.10
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
watchdog==6.0.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
